requirements/code.txt
Django
djangorestframework
djangorestframework-camel-case
djangorestframework-jwt
djangorestframework-xml
boto
django-filter
django-haystack
django-import-export
django-redactoreditor
pylibmc
redis
Jinja2
weibo
django-s3direct
pycrypto
django-ipware
django-simple-captcha
django-ckeditor
